Uploaded the technical report and PPT file
This research project was based on Tableau. Tableau was used to produce different data visualizations. The PPT file of the project gives the abstract summary of the project. It contains all the details of the different visualizations used and why those visualizations have been chosen.

The Technical report gives a detailed summary of the entire project.